pyproject.toml
[tool.poetry]
name = "wg-gp-crypto"
version = "0.1.0"
description = ""
authors = ["Chaipat Srikajohnlap <gorufu666@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.11"
aiohttp = "^3.9.1"
pandas = "^2.1.4"
python-dotenv = "^1.0.0"
mysql-connector-python = "^8.2.0"
schedule = "^1.2.1"
requests = "^2.31.0"
arrow = "^1.3.0"
websockets = "^12.0"
yfinance = "^0.2.33"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
